MM-43342: More strict permissions on categories

+++ server/api/categories.go
@@ -89,6 +89,11 @@ func (a *API) handleCreateCategory(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if !a.permissions.HasPermissionToTeam(session.UserID, teamID, model.PermissionViewTeam) {
+		a.errorResponse(w, r, model.NewErrPermission("access denied to team"))
+		return
+	}
+
 	createdCategory, err := a.app.CreateCategory(&category)
 	if err != nil {
 		a.errorResponse(w, r, err)
@@ -182,6 +187,11 @@ func (a *API) handleUpdateCategory(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if !a.permissions.HasPermissionToTeam(session.UserID, teamID, model.PermissionViewTeam) {
+		a.errorResponse(w, r, model.NewErrPermission("access denied to team"))
+		return
+	}
+
 	updatedCategory, err := a.app.UpdateCategory(&category)
 	if err != nil {
 		a.errorResponse(w, r, err)
@@ -238,6 +248,11 @@ func (a *API) handleDeleteCategory(w http.ResponseWriter, r *http.Request) {
 	auditRec := a.makeAuditRecord(r, "deleteCategory", audit.Fail)
 	defer a.audit.LogRecord(audit.LevelModify, auditRec)
 
+	if !a.permissions.HasPermissionToTeam(session.UserID, teamID, model.PermissionViewTeam) {
+		a.errorResponse(w, r, model.NewErrPermission("access denied to team"))
+		return
+	}
+
 	deletedCategory, err := a.app.DeleteCategory(categoryID, userID, teamID)
 	if err != nil {
 		a.errorResponse(w, r, err)
@@ -291,6 +306,11 @@ func (a *API) handleGetUserCategoryBoards(w http.ResponseWriter, r *http.Request
 
 	auditRec := a.makeAuditRecord(r, "getUserCategoryBoards", audit.Fail)
 	defer a.audit.LogRecord(audit.LevelModify, auditRec)
+
+	if !a.permissions.HasPermissionToTeam(session.UserID, teamID, model.PermissionViewTeam) {
+		a.errorResponse(w, r, model.NewErrPermission("access denied to team"))
+		return
+	}
 
 	categoryBlocks, err := a.app.GetUserCategoryBoards(userID, teamID)
 	if err != nil {
@@ -353,6 +373,11 @@ func (a *API) handleUpdateCategoryBoard(w http.ResponseWriter, r *http.Request) 
 	ctx := r.Context()
 	session := ctx.Value(sessionContextKey).(*model.Session)
 	userID := session.UserID
+
+	if !a.permissions.HasPermissionToTeam(session.UserID, teamID, model.PermissionViewTeam) {
+		a.errorResponse(w, r, model.NewErrPermission("access denied to team"))
+		return
+	}
 
 	// TODO: Check the category and the team matches
 	err := a.app.AddUpdateUserCategoryBoard(teamID, userID, map[string]string{boardID: categoryID})
